move catalogue and roster to tabs in classroom view; handle tab routing;
