+++ client/src/views/Classroom/ActivityCatalogue/ActivityCatalogue.less
@@ -1,13 +1,4 @@
-@import "../../assets/style.less";
-
-#main-header {
-  color: #colors[text-secondary];
-  font-size: 3em;
-  font-weight: bolder;
-  margin-left: 10%;
-  margin-bottom: 3%;
-  text-align: left;
-}
+@import "../../../assets/style.less";
 
 #search {
   .ant-select {
