+++ client/src/views/Classroom/Classroom.js
@@ -1,20 +1,23 @@
 import React, {useEffect, useState} from "react"
-import { Tabs } from 'antd';
+import {Tabs} from 'antd';
 import "./Classroom.less"
 
 import NavBar from "../../components/NavBar/NavBar";
 import {getClassroom} from "../../Utils/requests";
 import {getToken} from "../../Utils/AuthRequests";
+import Roster from "./Roster/Roster";
+import ActivityCatalogue from "./ActivityCatalogue/ActivityCatalogue";
 
 const {TabPane} = Tabs;
 
 export default function Classroom(props) {
     const [classroom, setClassroom] = useState({})
-    const {handleLogout, history} = props;
+    const {handleLogout, history, selectedActivity} = props;
     const path = history.location.pathname.split('/');
-    const classroomId = path[path.length-1];
+    const classroomId = path[path.length - 1];
+    const tab = history.location.hash.substring(1)
 
-    useEffect( () => {
+    useEffect(() => {
         getClassroom(classroomId, getToken()).then(classroom => {
             setClassroom(classroom);
         });
@@ -24,15 +27,19 @@ export default function Classroom(props) {
         <div className="container">
             <NavBar handleLogout={handleLogout}/>
             <div id='main-header'>{classroom.name}</div>
-            <Tabs defaultActiveKey="1">
-                <TabPane tab="Home" key="1">
-                    Content of card tab 1
+            <Tabs
+                defaultActiveKey={tab ? tab : "home"}
+                onChange={key => history.push(`#${key}`)}
+            >
+                <TabPane tab="Home" key="home">
+                    Your classroom info
                 </TabPane>
-                <TabPane tab="Roster" key="2">
-                    Content of card tab 2
+                <TabPane tab="Roster" key="roster">
+                    <Roster history={history} handleLogout={handleLogout} classroomId={classroomId}/>
                 </TabPane>
-                <TabPane tab="Learning Standards" key="3">
-                    Content of card tab 3
+                <TabPane tab="Learning Standards" key="learning-standards">
+                    <ActivityCatalogue history={history} selectedActivity={selectedActivity}
+                                       handleLogout={handleLogout}/>
                 </TabPane>
             </Tabs>
         </div>
